requirements_cpu.txt
tensorflow==2.12.0
opencv-python==4.8.1.78
mtcnn==0.1.1
scipy==1.11.4
numpy==1.24.3
flask==3.0.0
werkzeug==3.0.1
imutils==0.5.4
waitress==2.1.2
Pillow==10.1.0
python-dotenv==1.0.0
